Move resiliency test cases
