+++ stdlib/http/src/test/java/org/ballerinalang/test/cachingclient/HttpCachingClientTest.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.ballerinalang.test.net.http;
+package org.ballerinalang.test.cachingclient;
 
 import io.netty.handler.codec.http.HttpHeaders;
 import org.ballerinalang.compiler.CompilerPhase;
