+++ stdlib/http/src/test/resources/test-src/websocket/fail_onBinary.bal
