+++ stdlib/http/src/test/resources/test-src/websocket/fail_onText_JSON.bal
