+++ stdlib/http/src/test/resources/test-src/websocket/invalid_resource.bal
