+++ stdlib/http/src/test/resources/test-src/websocket/resource_return.bal
