+++ stdlib/http/src/test/resources/testng.xml
@@ -36,6 +36,9 @@ under the License.
             <package name="org.ballerinalang.test.services.cors.*"/>
             <package name="org.ballerinalang.test.services.dispatching.*"/>
             <package name="org.ballerinalang.test.services.nativeimpl.*"/>
+            <package name="org.ballerinalang.test.resiliency.*"/>
+            <package name="org.ballerinalang.test.websocket.*"/>
+            <package name="org.ballerinalang.test.cachingclient.*"/>
         </packages>
     </test>
 </suite>
